requirements.txt
aiohttp>=3.10.0
beautifulsoup4>=4.12.0
fastapi>=0.115.0
huggingface-hub>=0.24.0
numpy>=1.26.0
python-dotenv>=1.0.1
qdrant-client>=1.11.0
sentence-transformers>=3.0.0
uvicorn[standard]>=0.30.0